requirements.txt
# Core ML
torch>=2.0
torchvision>=0.15
albumentations>=1.3
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
pillow>=10.0
tqdm>=4.65

# Visualization
matplotlib>=3.7
seaborn>=0.12
plotly>=5.14

# Streamlit & UI
streamlit>=1.28
streamlit-option-menu>=0.3
streamlit-extras>=0.3

# Utilities
opencv-python>=4.8
python-dotenv>=1.0
openpyxl>=3.1  # For Excel export
